Delete task error

+++ src/main/java/hello/dao/TaskDAOImpl.java
@@ -47,8 +47,17 @@ public class TaskDAOImpl implements TaskDAO {
     }
 
     @Override
-    public void deleteTask(int taskId) {
-        entityManager.remove(getTaskById(taskId));
+    public int deleteTask(int taskId) {
+        if(taskExists(taskId)){
+            entityManager.remove(getTaskById(taskId));
+            if(taskExists(taskId)){
+                return -2;
+            }
+            return 1;
+        }else{
+            return -1;
+        }
+
     }
 
     @Override
